@@ -13,7 +13,7 @@ class ReadStatus extends StatelessWidget {
       appBar: AppBar(
         title: Text(
           'STATUS',
-          style: Theme.of(context).textTheme.headline6.copyWith(
+          style: Theme.of(context).textTheme.headline6!.copyWith(
                 fontWeight: FontWeight.bold,
               ),
         ),
@@ -29,9 +29,9 @@ class ReadStatus extends StatelessWidget {
                 return Center(child: CircularProgressIndicator());
               } else {
                 return ListView.builder(
-                  itemCount: snapshot.data.docs.length,
+                  itemCount: snapshot.data!.docs.length,
                   itemBuilder: (context, index) {
-                    DocumentSnapshot _status = snapshot.data.docs[index];
+                    DocumentSnapshot _status = snapshot.data!.docs[index];
                     return Card(
                       child: ListTile(
                         onTap: () {
